Update of 26/07/22 part 2
I upload the applied changes, as well as a better file tree.
At the beginning I launched the composer command in the test folder for the creation of an event, but since I'm going to use composer for the other events it's better to initiate it at the root of the project.